Remove legacy config file helpers
Finish the cleanup after the rooted filesystem migration.

This removes the remaining legacy config-file helper path from
tedge_utils::paths and migrates the last callers off it.

Changes include:
- centralize mosquitto-owned config writes in bridge code
- migrate bridge save paths to the rooted TedgePaths API
- switch tedge-agent upload directory creation to tedge_utils file
helpers
- remove DraftFile, persist_tempfile and create_directories from
paths.rs

This leaves the series ready for a final enforcement commit that adds
clippy restrictions and fixes any remaining test/support code.
