+++ crates/core/tedge/src/bridge/common_mosquitto_config.rs
@@ -1,9 +1,6 @@
 use camino::Utf8PathBuf;
 use tedge_config::TEdgeConfig;
-use tedge_utils::paths::DraftFile;
 use tokio::io::AsyncWriteExt;
-
-use super::TEDGE_BRIDGE_CONF_DIR_PATH;
 
 const COMMON_MOSQUITTO_CONFIG_FILENAME: &str = "tedge-mosquitto.conf";
 
@@ -231,23 +228,11 @@ impl CommonMosquittoConfig {
         &self,
         tedge_config: &TEdgeConfig,
     ) -> Result<(), tedge_utils::paths::PathsError> {
-        let dir_path = tedge_config.root_dir().join(TEDGE_BRIDGE_CONF_DIR_PATH);
-
-        tedge_utils::paths::create_directories(dir_path)?;
-
-        let config_path = self.file_path(tedge_config);
-        let mut config_draft = DraftFile::new(config_path).await?.with_mode(0o644);
-        self.serialize(&mut config_draft).await?;
-        config_draft.persist().await?;
+        let mut contents = Vec::new();
+        self.serialize(&mut contents).await?;
+        super::write_mosquitto_owned_config(tedge_config, &self.config_file, &contents).await?;
 
         Ok(())
-    }
-
-    fn file_path(&self, tedge_config: &TEdgeConfig) -> Utf8PathBuf {
-        tedge_config
-            .root_dir()
-            .join(TEDGE_BRIDGE_CONF_DIR_PATH)
-            .join(&self.config_file)
     }
 }
